requirements.txt
fastapi
uvicorn[standard]
pyserial
numpy
scipy
websockets
